+++ CHANGELOG.md
@@ -8,6 +8,7 @@ TODO: Date
 
 * Use `#let ident` to declare `ident`.
 * Use `#do parser_function => binding` to parse `binding` via `parser_function`.
+* Use `#do let parser_function => binding` to declare `binding` and parse it via `parser_function`.
 
 ## 0.0.5